requirements.txt
langchain_openai
langchain-google-genai 
langchain_core
python-dotenv
streamlit
langchain_community